clean-polish-and-remove-cloudconfig
