--- .dockerignore.dist
@@ -1,4 +1,0 @@
-.gobuild
-cache
-bin-dist/mayu*.tar.gz
-bin-dist/cluster
