+++ README.md
@@ -1,13 +1,11 @@
 [![CircleCI](https://circleci.com/gh/giantswarm/mayu.svg?&style=shield)](https://circleci.com/gh/giantswarm/mayu)
 [![](https://godoc.org/github.com/giantswarm/mayu?status.svg)](http://godoc.org/github.com/giantswarm/mayu) [![](https://img.shields.io/docker/pulls/giantswarm/mayu.svg)](http://hub.docker.com/giantswarm/mayu)
 [![Go Report Card](https://goreportcard.com/badge/github.com/giantswarm/mayu)](https://goreportcard.com/report/github.com/giantswarm/mayu)
-![Docker Repository on Quay](https://quay.io/repository/giantswarm/mayu/status "Docker Repository on Quay")](https://quay.io/repository/giantswarm/mayu)
+[![Docker Repository on Quay](https://quay.io/repository/giantswarm/mayu/status "Docker Repository on Quay")](https://quay.io/repository/giantswarm/mayu)
 
 # Mayu
 Mayu provides a set of mechanisms to bootstrap PXE-enabled bare metal nodes
-that must follow a specific configuration with Container Linux. It sets up fleet
-meta-data, and patched versions of fleet, etcd, and docker when using
-[Yochu](https://github.com/giantswarm/yochu).
+that must follow a specific configuration with Container Linux. 
 
 ## Prerequisites
 
@@ -19,20 +17,11 @@ be configured to boot via ethernet by default and share a network segment with
 the management node, so they get the PXE boot data from the management node on
 DHCP request.
 
-You need `dnsmasq` (>= 2.75) installed. Make sure dnsmasq is not running via an
-init script or systemd after you installed it, as mayu starts its own
-dnsmasq. In case you don't want to care about this, you maybe want to use the
-docker image, where the dependency is built in. This eases the setup.
-
-Further, mayu will keep track of all changes to the cluster by making git
-commits. This is a feature for production systems and requires `git` (> 1.7.4)
-installed. Use the `-no-git` flag when starting mayu to turn this feature off.
-
 Developing Mayu requires the following tools to be installed.
 
- * `wget`
- * `go-bindata`
- * `cpio`
+* `wget
+* `go-bindata
+* `cpio`
 
 ## Getting Mayu
