+++ client/client_test.go
@@ -54,86 +54,6 @@ func newClientAndServer(t *testing.T, handler http.Handler) (*client.Client, *ht
 }
 
 //
-// Client.SetMetadata
-//
-
-// Test_Client_001 checks for Client.SetMetadata to provide proper information
-// to the server as expected.
-func Test_Client_001(t *testing.T) {
-	var response testResponse
-
-	newClient, ts := newClientAndServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		body, err := ioutil.ReadAll(r.Body)
-		r.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		response = testResponse{
-			Body:   body,
-			Header: r.Header,
-			Method: r.Method,
-			Path:   r.URL.Path,
-		}
-
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("ok"))
-	}))
-	defer ts.Close()
-
-	err := newClient.SetMetadata("serial", "key1=value1,key2=value2")
-	if err != nil {
-		t.Fatalf("Client.SetMetadata returned error: %#v", err)
-	}
-
-	data, err := json.Marshal(hostmgr.Host{
-		FleetMetadata: []string{"key1=value1", "key2=value2"},
-	})
-	if err != nil {
-		t.Fatalf("json.Marshal returned error: %#v", err)
-	}
-	if string(response.Body) != string(data) {
-		t.Fatalf("expected response body to be '%s', got '%s'", string(response.Body), string(data))
-	}
-
-	assertHeader(t, response, "content-type", []string{"application/json"})
-	assertMethod(t, response, "PUT")
-	assertPath(t, response, fmt.Sprintf("/admin/host/%s/set_metadata", "serial"))
-}
-
-// Test_Client_002 checks for Client.SetMetadata to provide proper error
-// information to the client as expected, when there are errors returned from
-// the server.
-func Test_Client_002(t *testing.T) {
-	newClient, ts := newClientAndServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("internal server error"))
-	}))
-	defer ts.Close()
-
-	err := newClient.SetMetadata("serial", "key1=value1,key2=value2")
-	if err == nil {
-		t.Fatalf("Client.SetMetadata NOT returned error")
-	}
-}
-
-// Test_Client_003 checks for Client.SetMetadata to provide proper error
-// information to the client as expected, when there is no server running.
-func Test_Client_003(t *testing.T) {
-	newClient, ts := newClientAndServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("internal server error"))
-	}))
-	// Immediatelly close the server.
-	ts.Close()
-
-	err := newClient.SetMetadata("serial", "key1=value1,key2=value2")
-	if err == nil {
-		t.Fatalf("Client.SetMetadata NOT returned error")
-	}
-}
-
-//
 // Client.SetProviderId
 //
 
@@ -290,86 +210,6 @@ func Test_Client_009(t *testing.T) {
 	err := newClient.SetIPMIAddr("serial", "127.0.0.1")
 	if err == nil {
 		t.Fatalf("Client.SetIPMIAddr NOT returned error")
-	}
-}
-
-//
-// Client.SetCabinet
-//
-
-// Test_Client_010 checks for Client.SetCabinet to provide proper information
-// to the server as expected.
-func Test_Client_010(t *testing.T) {
-	var response testResponse
-
-	newClient, ts := newClientAndServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		body, err := ioutil.ReadAll(r.Body)
-		r.Body.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		response = testResponse{
-			Body:   body,
-			Header: r.Header,
-			Method: r.Method,
-			Path:   r.URL.Path,
-		}
-
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("ok"))
-	}))
-	defer ts.Close()
-
-	err := newClient.SetCabinet("serial", "101")
-	if err != nil {
-		t.Fatalf("Client.SetCabinet returned error: %#v", err)
-	}
-
-	data, err := json.Marshal(hostmgr.Host{
-		Cabinet: uint(101),
-	})
-	if err != nil {
-		t.Fatalf("json.Marshal returned error: %#v", err)
-	}
-	if string(response.Body) != string(data) {
-		t.Fatalf("expected response body to be '%s', got '%s'", string(response.Body), string(data))
-	}
-
-	assertHeader(t, response, "content-type", []string{"application/json"})
-	assertMethod(t, response, "PUT")
-	assertPath(t, response, fmt.Sprintf("/admin/host/%s/set_cabinet", "serial"))
-}
-
-// Test_Client_011 checks for Client.SetCabinet to provide proper error
-// information to the client as expected, when there are errors returned from
-// the server.
-func Test_Client_011(t *testing.T) {
-	newClient, ts := newClientAndServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("internal server error"))
-	}))
-	defer ts.Close()
-
-	err := newClient.SetCabinet("serial", "101")
-	if err == nil {
-		t.Fatalf("Client.SetCabinet NOT returned error")
-	}
-}
-
-// Test_Client_012 checks for Client.SetCabinet to provide proper error
-// information to the client as expected, when there is no server running.
-func Test_Client_012(t *testing.T) {
-	newClient, ts := newClientAndServer(t, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("internal server error"))
-	}))
-	// Immediatelly close the server.
-	ts.Close()
-
-	err := newClient.SetCabinet("serial", "101")
-	if err == nil {
-		t.Fatalf("Client.SetCabinet NOT returned error")
 	}
 }
