+++ files/conf/lacp-bonding.conf
@@ -1,0 +1,1 @@
+options bonding miimon=100 mode=4 lacp_rate=1
