+++ files/conf/module-bonding.conf
@@ -1,0 +1,1 @@
+bonding
