+++ files/my-service/my-service.conf
@@ -1,0 +1,2 @@
+[section1]
+value1=test
