+++ hostmgr/host.go
@@ -15,39 +15,25 @@ const hostConfFile = "conf.json"
 
 // Host represents a node within the mayu cluster.
 type Host struct {
-	Id                 int       `json:",omitempty"`
-	ProviderId         string    `json:",omitempty"`
-	Enabled            bool      `json:",omitempty"`
-	Name               string    `json:",omitempty"`
-	Serial             string    `json:",omitempty"`
-	MacAddresses       []string  `json:",omitempty"`
-	InternalAddr       net.IP    `json:",omitempty"`
-	BondInterfaces     []string  `json:",omitempty"`
-	Cabinet            uint      `json:",omitempty"`
-	MachineOnCabinet   uint      `json:",omitempty"`
-	IPMIAddr           net.IP    `json:",omitempty"`
-	Hostname           string    `json:",omitempty"`
-	MachineID          string    `json:",omitempty"`
-	ConnectedNIC       string    `json:",omitempty"`
-	FleetMetadata      FleetMeta `json:",omitempty"`
-	FleetDisableEngine bool      `json:",omitempty"`
-	KeepDiskData       bool      `json:",omitempty"`
-	LastBoot           time.Time `json:",omitempty"`
-	Profile            string    `json:",omitempty"`
-	EtcdClusterToken   string    `json:",omitempty"`
+	Id               int       `json:",omitempty"`
+	ProviderId       string    `json:",omitempty"`
+	Enabled          bool      `json:",omitempty"`
+	Name             string    `json:",omitempty"`
+	Serial           string    `json:",omitempty"`
+	MacAddresses     []string  `json:",omitempty"`
+	InternalAddr     net.IP    `json:",omitempty"`
+	IPMIAddr         net.IP    `json:",omitempty"`
+	Hostname         string    `json:",omitempty"`
+	MachineID        string    `json:",omitempty"`
+	LastBoot         time.Time `json:",omitempty"`
+	Profile          string    `json:",omitempty"`
+	EtcdClusterToken string    `json:",omitempty"`
 
 	Overrides map[string]interface{} `json:",omitempty"`
 
 	State hostState
 
 	CoreOSVersion string `json:",omitempty"`
-	MayuVersion   string `json:",omitempty"`
-	DockerVersion string `json:",omitempty"`
-	FleetVersion  string `json:",omitempty"`
-	EtcdVersion   string `json:",omitempty"`
-	RktVersion    string `json:",omitempty"`
-	K8sVersion    string `json:",omitempty"`
-	YochuVersion  string `json:",omitempty"`
 
 	hostDir     *os.File
 	lastModTime time.Time
@@ -60,23 +46,19 @@ type IPMac struct {
 	MacAddr string
 }
 
-func (fleetMeta FleetMeta) Has(m string) bool {
-	for _, v := range fleetMeta {
-		if strings.ToLower(m) == strings.ToLower(v) {
-			return true
-		}
-	}
-	return false
-}
-
-func (fleetMeta FleetMeta) String() string {
-	return strings.Join(fleetMeta, ",")
-}
-
 // Commit stores the given msg in git version control.
 func (h *Host) Commit(msg string) error {
 	h.save()
 	return h.maybeGitCommit(h.Serial + ": " + msg)
+}
+
+func genMachineID() string {
+	b := make([]byte, 16)
+	_, err := rand.Read(b)
+	if err != nil {
+		panic(err)
+	}
+	return hex.EncodeToString(b)
 }
 
 // HostFromDir takes a path to a host directory within the cluster directory
@@ -102,18 +84,6 @@ func HostFromDir(hostdir string) (*Host, error) {
 	h.lastModTime = fi.ModTime()
 
 	return h, nil
-}
-
-func genMachineID(cabinet, machineOnCabinet uint) string {
-	b := make([]byte, 16)
-	_, err := rand.Read(b)
-	if err != nil {
-		panic(err)
-	}
-	b[0] = byte(0xff & cabinet)
-	b[1] = byte(0xff & machineOnCabinet)
-
-	return hex.EncodeToString(b)
 }
 
 func createHost(serial string, hostDir string) (*Host, error) {
