--- pxemgr/cloudconfig.go
@@ -1,84 +1,0 @@
-package pxemgr
-
-import (
-	"fmt"
-	"io"
-	"io/ioutil"
-	"os"
-	"path"
-	"text/template"
-
-	"github.com/giantswarm/mayu/hostmgr"
-	"github.com/golang/glog"
-)
-
-var snippetsFiles []string
-
-func maybeInitSnippets(snippets string) {
-	if snippetsFiles != nil {
-		return
-	}
-	snippetsFiles = []string{}
-
-	if len(snippets) > 0 {
-		if _, err := os.Stat(snippets); err == nil {
-			if fis, err := ioutil.ReadDir(snippets); err == nil {
-				for _, fi := range fis {
-					snippetsFiles = append(snippetsFiles, path.Join(snippets, fi.Name()))
-				}
-			}
-		}
-	}
-}
-
-func getTemplate(path, snippets string) (*template.Template, error) {
-	maybeInitSnippets(snippets)
-	templates := []string{path}
-	templates = append(templates, snippetsFiles...)
-	glog.V(10).Infof("templates: %+v\n", templates)
-
-	return template.ParseFiles(templates...)
-}
-
-func (mgr *pxeManagerT) WriteLastStageCC(host hostmgr.Host, wr io.Writer) error {
-	etcdClusterToken := mgr.cluster.Config.DefaultEtcdClusterToken
-
-	if host.EtcdClusterToken != "" {
-		etcdClusterToken = host.EtcdClusterToken
-	}
-
-	mergedTemplatesEnv := mgr.config.TemplatesEnv
-	for k, v := range host.Overrides {
-		mergedTemplatesEnv[k] = v
-	}
-
-	ctx := struct {
-		Host             hostmgr.Host
-		EtcdDiscoveryUrl string
-		ClusterNetwork   Network
-		MayuHost         string
-		MayuPort         int
-		MayuURL          string
-		PostBootURL      string
-		NoTLS            bool
-		TemplatesEnv     map[string]interface{}
-	}{
-		Host:             host,
-		ClusterNetwork:   mgr.config.Network,
-		EtcdDiscoveryUrl: fmt.Sprintf("%s/%s", mgr.etcdDiscoveryUrl, etcdClusterToken),
-		MayuHost:         mgr.config.Network.BindAddr,
-		MayuPort:         mgr.apiPort,
-		MayuURL:          mgr.apiURL(),
-		PostBootURL:      mgr.apiURL() + "/admin/host/" + host.Serial + "/boot_complete",
-		NoTLS:            mgr.noTLS,
-		TemplatesEnv:     mergedTemplatesEnv,
-	}
-
-	tmpl, err := getTemplate(mgr.lastStageCloudconfig, mgr.templateSnippets)
-	if err != nil {
-		glog.Fatalln(err)
-		return err
-	}
-
-	return tmpl.Execute(wr, ctx)
-}
