--- template_snippets/cloudconfig/core_snippet.yaml
@@ -1,29 +1,0 @@
-{{define "core_units"}}
-  - name: my-core-unit.service
-    enable: true
-    command: start
-    content: |
-      [Unit]
-      Description=my unit Service
-
-      [Service]
-      Restart=on-failure
-      RestartSec=1
-      StartLimitInterval=300s
-      StartLimitBurst=3
-      TimeoutStartSec=0
-      ExecStart=/bin/echo my-unit-running-command
-      ExecStop=/bin/echo my-unit-stop-command
-
-{{end}}
-
-{{define "core_files"}}
-- path: "/etc/my-conf-file"
-  owner: root
-  permissions: 644
-  content: |
-    asd=1
-    value2=3
-    some_other_stuff=true
-    and_so_on="string value"
-{{end}}
