--- template_snippets/cloudconfig/net_bond.yaml
@@ -1,62 +1,0 @@
-{{define "net_bond"}}
-  - name: 10-bond0-slave.network
-    runtime: true
-    content: |
-      [Match]
-      Name=en*
-
-      [Network]
-      Bond=bond0
-
-  - name: 20-bond0.network
-    runtime: true
-    content: |
-      [Match]
-      Name=bond0
-
-      [Network]
-      DHCP=none
-      VLAN=bond0.{{.ClusterNetwork.VlanId}}
-
-  - name: 15-bond0.netdev
-    runtime: true
-    content: |
-      [NetDev]
-      Name=bond0
-      Kind=bond
-
-      [Bond]
-{{- if eq .ClusterNetwork.BondMode "lacp" }}
-      Mode=802.3ad
-      LACPTransmitRate=fast
-      MIIMonitorSec=1s
-      UpDelaySec=3s
-      DownDelaySec=9s
-{{else}}
-      Mode=balance-rr
-{{end}
-  - name: 35-bond0.{{.ClusterNetwork.VlanId}}.netdev
-    runtime: true
-    content: |
-      [NetDev]
-      Name=bond0.{{.ClusterNetwork.VlanId}}
-      Kind=vlan
-
-      [VLAN]
-      Id={{.ClusterNetwork.VlanId}}
-
-  - name: 40-vlan.network
-    runtime: true
-    content: |
-      [Match]
-      Name=bond0.{{.ClusterNetwork.VlanId}}
-
-      [Network]
-      Address={{.Host.InternalAddr}}/{{.ClusterNetwork.SubnetSize}}
-      Gateway={{.ClusterNetwork.SubnetGateway}}
-      {{ range $server := .ClusterNetwork.DNS }}DNS={{ $server }}
-      {{ end }}
-      {{ range $server := .ClusterNetwork.NTP }}NTP={{ $server }}
-      {{ end }}
-
-{{end}}
