--- template_snippets/cloudconfig/worker_snippet.yaml
@@ -1,27 +1,0 @@
-{{define "worker_units"}}
-  - name: my-worker-unit.service
-    enable: true
-    command: start
-    content: |
-      [Unit]
-      Description=my unit Service
-
-      [Service]
-      Restart=on-failure
-      RestartSec=1
-      StartLimitInterval=300s
-      StartLimitBurst=3
-      TimeoutStartSec=0
-      ExecStart=/bin/echo do something
-      ExecStop=/bin/echo stop something
-
-{{end}}
-
-{{define "worker_files"}}
-- path: "/tmp/log-file"
-  owner: root
-  permissions: 644
-  content: |
-    some content
-    some other content
-{{end}}
