--- template_snippets/ignition/net_singlenic.yaml
@@ -1,23 +1,0 @@
-{{define "net_singlenic"}}
-networkd:
-  units:
-  - name: 10-nodhcp.network
-    contents: |
-      [Match]
-      Name=*
-
-      [Network]
-      DHCP=no
-  - name: 00-{{.Host.ConnectedNIC}}.network
-    contents: |
-      [Match]
-      Name={{.Host.ConnectedNIC}}
-
-      [Network]
-      Address={{.Host.InternalAddr}}/{{.ClusterNetwork.SubnetSize}}
-      Gateway={{.ClusterNetwork.SubnetGateway}}
-      {{ range $server := .ClusterNetwork.DNS }}DNS={{ $server }}
-      {{ end }}
-      {{ range $server := .ClusterNetwork.NTP }}NTP={{ $server }}
-      {{ end }}
-{{end}}
